Refactor chart component by extracting hooks, themes, and utils
