+++ src/components/ui/button.tsx
@@ -26,4 +26,4 @@ function Button({
   );
 }
 
-export { Button, buttonVariants };
+export { Button };
