+++ src/components/ui/chart-hooks.ts
@@ -1,0 +1,20 @@
+import * as React from "react";
+import type { ChartConfig } from "./chart";
+
+type ChartContextProps = {
+  config: ChartConfig;
+};
+
+const ChartContext = React.createContext<ChartContextProps>({
+  config: {},
+});
+
+export function useChart() {
+  const context = React.useContext(ChartContext);
+  if (!context) {
+    throw new Error("useChart must be used within a ChartProvider");
+  }
+  return context;
+}
+
+export { ChartContext };
