+++ src/components/ui/chart-themes.ts
@@ -1,0 +1,40 @@
+export const THEMES = {
+  light: {
+    background: "hsl(0 0% 100%)",
+    foreground: "hsl(222.2 84% 4.9%)",
+    primary: "hsl(222.2 47.4% 11.2%)",
+    secondary: "hsl(210 40% 96%)",
+    muted: "hsl(210 40% 96%)",
+    accent: "hsl(210 40% 96%)",
+    destructive: "hsl(0 84.2% 60.2%)",
+    border: "hsl(214.3 31.8% 91.4%)",
+    input: "hsl(214.3 31.8% 91.4%)",
+    ring: "hsl(222.2 84% 4.9%)",
+    chart: {
+      "1": "hsl(222.2 84% 4.9%)",
+      "2": "hsl(210 40% 96%)",
+      "3": "hsl(210 40% 96%)",
+      "4": "hsl(210 40% 96%)",
+      "5": "hsl(210 40% 96%)",
+    },
+  },
+  dark: {
+    background: "hsl(222.2 84% 4.9%)",
+    foreground: "hsl(210 40% 98%)",
+    primary: "hsl(210 40% 98%)",
+    secondary: "hsl(217.2 32.6% 17.5%)",
+    muted: "hsl(217.2 32.6% 17.5%)",
+    accent: "hsl(217.2 32.6% 17.5%)",
+    destructive: "hsl(0 62.8% 30.6%)",
+    border: "hsl(217.2 32.6% 17.5%)",
+    input: "hsl(217.2 32.6% 17.5%)",
+    ring: "hsl(212.7 26.8% 83.9%)",
+    chart: {
+      "1": "hsl(210 40% 98%)",
+      "2": "hsl(217.2 32.6% 17.5%)",
+      "3": "hsl(217.2 32.6% 17.5%)",
+      "4": "hsl(217.2 32.6% 17.5%)",
+      "5": "hsl(217.2 32.6% 17.5%)",
+    },
+  },
+} as const;
