+++ src/components/ui/toggle.tsx
@@ -23,4 +23,4 @@ function Toggle({
   );
 }
 
-export { Toggle, toggleVariants };
+export { Toggle };
